requirements.txt
py-cord~=2.5
jishaku~=2.5
rich~=13.7
tortoise-orm[accel]~=0.20
